Issue: 44
M      src/main/java/com/sun/faces/extensions/avatar/renderkit/AjaxZoneRenderer.java

First problem: there was a bug in getInteractionType where it was always
returning "input" even if there were no input components.  This was
due to including the zone itself in the search.  Because the zone itself
is an ActionSource, getInteractionType was returning input.

This change excludes the zone from the instanceof test.

svn path=/trunk/; revision=348