Alternate fix for SPEC-323. r=edburns
This is an alternate proposal to 323.  Instead of having the
logic for setValue(Object) check the javax.faces.INTERPRET_EMPTY_STRING_SUBMITTED_VALUES_AS_NULL
via a FacesContext ThreadLocal lookup, move this logic to 
UIInput.validate() instead.

Doing so here solves three issues.
  1.  Having an additional FacesContext thread local lookup
  2.  Having zero-length Strings pushed to the model 
  3.  Prevent ValueChangeListeners firing when they shouldn't.

Issues 1 and 2 should already be clear, so I'll touch on issue 
3 which was brought up by SAP about a month and a half ago.

Consider:

<h:form id="form">
    <h:inputText id="string"
                 value="#{bean.stringProperty}"
                 valueChangeListener="#{bean.valueChange}"/>
</h:form>

Where #{bean.stringProperty} is initially null.  In JSF 1.2, if this
form is submitted without entering any data into the input control,
the ValueChangeListener is triggered when in reality, it shouldn't have
been.

So we resolve this and the other two issues by adding the following
block of code:

         Object submittedValue = getSubmittedValue();
         if (submittedValue == null) {
             return;
         }
 
+        // If non-null, an instanceof String, and we're configured to treat
+        // zero-length Strings as null:
+        //   call setSubmittedValue(null)
+        if ((considerEmptyStringNull(context)
+             && submittedValue instanceof String 
+             && ((String) submittedValue).length() == 0)) {
+            setSubmittedValue(null);
+            submittedValue = null;
+        }
+

If getSubmittedValue() returns null, simply return.

If the javax.faces.INTERPRET_EMPTY_STRING_SUBMITTED_VALUES_AS_NULL
option is enabled, and the submitted value is an empty string,
set the local submittedValue to null and call setSubmittedValue(null).
And continue processing.  Validation won't occur, a VCE won't be triggers
since both the old and new values are null, and the model will get a null
value.

svn path=/trunk/; revision=556